+++ data_migrator/models/__init__.py
@@ -12,8 +12,3 @@ This module contains all classes for models, managers and fields
 from .base import Model
 from .manager import SimpleManager
 from .fields import IntField, NullIntField, StringField, NullStringField, BooleanField, UUIDField, NullField, JSONField, MappingField, HiddenField
-
-
-__all__ = [
-    'HiddenField'
-]